+++ packages/js/src/settings/hocs/index.js
@@ -1,1 +1,2 @@
 export { default as withDisabledMessageSupport } from "./with-disabled-message-support";
+export { default as withFormikError } from "./with-formik-error";
